Added gyroscope and an easier way to use the dashboard
-Last year and this year we used a special gyroscope that functions
differently than other sensors.  It is called the navX MXP and it
requires an extra class base to use on an FRC robot.  The documentation
can be found here http://www.pdocs.kauailabs.com/navx-mxp/
-I added an interface called DashboardSender which most subsytems will
implement.  This is just to make it easier for me to easily send data to
the SmartDashboard or regular dashboard.
-This year, I will be using the SmartDashboard for testing, and regular
(LabView) dashboard for competitions.