Reorganized classes, updated vision (again I know)
-Overall I just renamed some classes so they are more organized, as well
as moved all of the vision related commands to a separate package than
the other commands.
-Vision now actually tracks distance and shoots based on that distance,
finally.  For real this time I swear to God.
-Added a rotate command for vision that continuously accesses the gamma
variable and rotates based on readings from the Vision class rather than
calculate and angle for the robot to turn and make the gyro do all of
the work.